+++ GreenFlamingosApp/GreenFlamingosWebApp/Models/Shot.cs
@@ -13,7 +13,7 @@ namespace GreenFlamingosWebApp.Models
         {
             DrinkType = "Shot";
         }
-        public Shot(string name, User owner, string mainIgredient, int capacity, int alcoholContent, int calories, List<string> ingriedients, string description, List<string> preparation, string imageUrl)
+        public Shot(string name, User owner, string mainIgredient, int capacity, float alcoholContent, int calories, List<string> ingriedients, string description, List<string> preparation, string imageUrl)
         {
             DrinkType = "Shot";
             Name = name;
